allowing the api url to be set as an option
