+++ lib/opentok.rb
@@ -19,7 +19,8 @@ module OpenTok
 
   VERSION = "tbrb-v0.91.2011-02-17"
   API_URL = "https://staging.tokbox.com/hl"
-
+  
+  require 'monkey_patches'
   require 'open_tok/exceptions'
   require 'open_tok/open_tok_sdk'
 end
